opti

+++ src/main/java/cn/nukkit/level/generator/Nether.java
@@ -20,14 +20,14 @@ import java.util.*;
 
 public class Nether extends Generator {
 
-    private static double coordinateScale = (double) getConfig("nether.coordinate-scale", 684.412d);
-    private static double heightScale = (double) getConfig("nether.height.scale", 2053.236d);
-    private static double heightNoiseScaleX = (double) getConfig("nether.height.noise-scale.x", 100.0d); // depthNoiseScaleX
-    private static double heightNoiseScaleZ = (double) getConfig("nether.height.noise-scale.x", 100.0d); // depthNoiseScaleZ
-    private static double detailNoiseScaleX = (double) getConfig("nether.detail.noise-scale.x", 80.0d);  // mainNoiseScaleX
-    private static double detailNoiseScaleY = (double) getConfig("nether.detail.noise-scale.y", 60.0d);  // mainNoiseScaleY
-    private static double detailNoiseScaleZ = (double) getConfig("nether.detail.noise-scale.z", 80.0d);  // mainNoiseScaleZ
-    private static double surfaceScale = (double) getConfig("nether.surface-scale", 0.0625d);
+    private static double coordinateScale = getConfig("nether.coordinate-scale", 684.412d);
+    private static double heightScale = getConfig("nether.height.scale", 2053.236d);
+    private static double heightNoiseScaleX = getConfig("nether.height.noise-scale.x", 100.0d); // depthNoiseScaleX
+    private static double heightNoiseScaleZ = getConfig("nether.height.noise-scale.x", 100.0d); // depthNoiseScaleZ
+    private static double detailNoiseScaleX = getConfig("nether.detail.noise-scale.x", 80.0d);  // mainNoiseScaleX
+    private static double detailNoiseScaleY = getConfig("nether.detail.noise-scale.y", 60.0d);  // mainNoiseScaleY
+    private static double detailNoiseScaleZ = getConfig("nether.detail.noise-scale.z", 80.0d);  // mainNoiseScaleZ
+    private static double surfaceScale = getConfig("nether.surface-scale", 0.0625d);
 
     private final Map<String, Map<String, OctaveGenerator>> octaveCache = new HashMap<String, Map<String, OctaveGenerator>>();
     private final double[][][] density = new double[5][5][17];
